Feature/ac 961

+++ server/api/tagger/dao/collection-dao.js
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016.
+ * Copyright (c) 2017.
  *
  *     This program is free software: you can redistribute it and/or modify
  *     it under the terms of the GNU General Public License as published by
@@ -21,8 +21,25 @@
 
 'use strict';
 
-const taggerSchema = require('../models/index');
+const taggerSchema = require('../schema/index');
+const utils = require('./utils');
+const logger = require('../utils/error-logger');
+const path = require('path');
+const filename = path.basename(__filename);
+const paramErrorMessage = 'A parameter for a collection query is not defined.';
 const taggerDao = {};
+
+/**
+ * Returns 500 error for missing parameter. This error is thrown
+ * before the dao promise is returned.
+ * @returns {Error}
+ * @private
+ */
+function _errorResponse() {
+  let error = new Error('Error: missing query parameter - ' + filename);
+  error.status = 500;
+  return error;
+}
 
 taggerDao.retrieveAllCollections = () => {
 
@@ -32,7 +49,23 @@ taggerDao.retrieveAllCollections = () => {
 
 };
 
+taggerDao.retrieveAllPublishedCollections = () => {
+  return taggerSchema.Collection.findAll({
+    where: {
+      published: true
+    },
+    order: [['title', 'ASC']]
+  });
+
+
+};
+
 taggerDao.countCTypesByArea = (areaId) => {
+
+  if (!areaId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
 
   return taggerSchema.sequelize.query('SELECT ctype, COUNT(*) as count FROM AreaTargets ' +
     'LEFT JOIN Collections ON AreaTargets.CollectionId = Collections.id ' +
@@ -46,6 +79,11 @@ taggerDao.countCTypesByArea = (areaId) => {
 
 taggerDao.browseTypesByArea = (areaId) => {
 
+  if (!areaId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
   return taggerSchema.sequelize.query('SELECT Collections.browseType, COUNT(Collections.id) as count from AreaTargets ' +
     'join Collections on AreaTargets.CollectionId=Collections.id where AreaTargets.AreaId = ? group by Collections.browseType',
     {
@@ -56,6 +94,11 @@ taggerDao.browseTypesByArea = (areaId) => {
 };
 
 taggerDao.repoTypesByArea = (areaId) => {
+
+  if (!areaId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
 
   return taggerSchema.sequelize.query('SELECT repoType, COUNT(*) as count FROM AreaTargets ' +
     'LEFT JOIN Collections ON AreaTargets.CollectionId = Collections.id ' +
@@ -68,6 +111,11 @@ taggerDao.repoTypesByArea = (areaId) => {
 };
 
 taggerDao.findCollectionsInArea = (areaId) => {
+
+  if (!areaId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
 
   return taggerSchema.AreaTarget.findAll({
     where: {
@@ -82,6 +130,11 @@ taggerDao.findCollectionsInArea = (areaId) => {
 
 taggerDao.findAreasForCollection = (collId) => {
 
+  if (!collId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
   return taggerSchema.AreaTarget.findAll({
     where: {
       CollectionId: collId
@@ -91,6 +144,11 @@ taggerDao.findAreasForCollection = (collId) => {
 };
 
 taggerDao.findItemContentTarget = (collId, typeId) => {
+
+  if (!collId || !typeId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
 
   return taggerSchema.ItemContentTarget.find(
     {
@@ -104,6 +162,11 @@ taggerDao.findItemContentTarget = (collId, typeId) => {
 
 taggerDao.createItemContentTarget = (collId, typeId) => {
 
+  if (!collId || !typeId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
   return taggerSchema.ItemContentTarget.create(
     {
       CollectionId: collId,
@@ -114,6 +177,11 @@ taggerDao.createItemContentTarget = (collId, typeId) => {
 };
 
 taggerDao.deleteItemContentTarget = (collId, typeId) => {
+
+  if (!collId || !typeId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
 
   return taggerSchema.ItemContentTarget.destroy(
     {
@@ -128,6 +196,11 @@ taggerDao.deleteItemContentTarget = (collId, typeId) => {
 
 taggerDao.findContentTypesForCollection = (collId) => {
 
+  if (!collId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
   return taggerSchema.ItemContentTarget.findAll(
     {
       where: {
@@ -141,6 +214,12 @@ taggerDao.findContentTypesForCollection = (collId) => {
 
 taggerDao.checkForExistingTagTarget = (collId, tagId) => {
 
+  if (!collId || !tagId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
+
   return taggerSchema.TagTarget.find(
     {
       where: {
@@ -153,6 +232,11 @@ taggerDao.checkForExistingTagTarget = (collId, tagId) => {
 
 taggerDao.addTagTarget = (collId, tagId) => {
 
+  if (!collId || !tagId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
   return taggerSchema.TagTarget.create(
     {
       CollectionId: collId,
@@ -163,6 +247,11 @@ taggerDao.addTagTarget = (collId, tagId) => {
 };
 
 taggerDao.deleteTagTarget = (collId, tagId) => {
+
+  if (!collId || !tagId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
 
   return taggerSchema.TagTarget.destroy(
     {
@@ -177,6 +266,11 @@ taggerDao.deleteTagTarget = (collId, tagId) => {
 
 taggerDao.addCollectionToArea = (collId, areaId) => {
 
+  if (!collId || !areaId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
   return taggerSchema.AreaTarget.create(
     {
       CollectionId: collId,
@@ -186,6 +280,12 @@ taggerDao.addCollectionToArea = (collId, areaId) => {
 };
 
 taggerDao.getAreaIdsForCollection = (collId) => {
+
+  if (!collId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
 
   return taggerSchema.AreaTarget.findAll(
     {
@@ -200,6 +300,12 @@ taggerDao.getAreaIdsForCollection = (collId) => {
 
 taggerDao.checkAreaAssociation = (collId, areaId) => {
 
+  if (!collId || !areaId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
+
   return taggerSchema.AreaTarget.find(
     {
       where: {
@@ -211,6 +317,11 @@ taggerDao.checkAreaAssociation = (collId, areaId) => {
 };
 
 taggerDao.removeCollectionFromArea = (areaId, collId) => {
+
+  if (!collId || !areaId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
 
   return taggerSchema.AreaTarget.destroy({
     where: {
@@ -224,6 +335,12 @@ taggerDao.removeCollectionFromArea = (areaId, collId) => {
 
 taggerDao.findCollectionById = (collId) => {
 
+  if (!collId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
+
   return taggerSchema.Collection.find(
     {
       where: {
@@ -233,6 +350,12 @@ taggerDao.findCollectionById = (collId) => {
 };
 
 taggerDao.findCategoryAssociation = (collId) => {
+
+  if (!collId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
 
   return taggerSchema.CategoryTarget.find(
     {
@@ -244,6 +367,12 @@ taggerDao.findCategoryAssociation = (collId) => {
 };
 
 taggerDao.getCategoryForCollection = (collId) => {
+
+  if (!collId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
 
   return taggerSchema.CategoryTarget.find(
     {
@@ -257,6 +386,12 @@ taggerDao.getCategoryForCollection = (collId) => {
 
 taggerDao.setPublicationStatus = (status, collId) => {
 
+  if (!collId || !status) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
+
   return taggerSchema.Collection.update({
       published: status
     },
@@ -269,6 +404,12 @@ taggerDao.setPublicationStatus = (status, collId) => {
 
 taggerDao.getPublicationStatus = (collId) => {
 
+  if (!collId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
+
   return taggerSchema.Collection.find(
     {
       where: {
@@ -279,6 +420,12 @@ taggerDao.getPublicationStatus = (collId) => {
 };
 
 taggerDao.updateCollection = (update, id) => {
+
+  if (!update || !id) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
 
   return taggerSchema.Collection.update(update,
     {
@@ -291,11 +438,22 @@ taggerDao.updateCollection = (update, id) => {
 
 taggerDao.addCollectionToCategory = (id, category) => {
 
+  if (!id || !category) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
   return taggerSchema.CategoryTarget.create({CollectionId: id, CategoryId: category});
 
 };
 
 taggerDao.updateCollectionCategory = (id, category) => {
+
+  if (!id || !category) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
 
   return taggerSchema.CategoryTarget.update({
       CategoryId: category
@@ -309,16 +467,25 @@ taggerDao.updateCollectionCategory = (id, category) => {
 };
 
 taggerDao.deleteCategoryFromCollection = (id, category) => {
+  if (!id || !category) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
 
- return taggerSchema.CategoryTarget.destroy({
-   where: {
-     CategoryId: category,
-     CollectionId: id
-   }
- });
+  return taggerSchema.CategoryTarget.destroy({
+    where: {
+      CategoryId: category,
+      CollectionId: id
+    }
+  });
 };
 
 taggerDao.deleteCollection = (collId) => {
+
+  if (!collId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
 
   return taggerSchema.Collection.destroy({
     where: {
@@ -330,6 +497,11 @@ taggerDao.deleteCollection = (collId) => {
 
 taggerDao.addNewCollection = (title, browseType, repoType, ctype) => {
 
+  if (!title || !browseType || !repoType || !ctype) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
   return taggerSchema.Collection.create({
     title: title,
     browseType: browseType,
@@ -340,6 +512,11 @@ taggerDao.addNewCollection = (title, browseType, repoType, ctype) => {
 };
 
 taggerDao.updateCollectionImage = (collId, imageName) => {
+
+  if (!collId || !imageName) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
 
   return taggerSchema.Collection.update(
     {
@@ -354,24 +531,20 @@ taggerDao.updateCollectionImage = (collId, imageName) => {
 
 };
 
-taggerDao.findTagsForCollection = (collId) => {
-
-  return taggerSchema.TagTarget.findAll(
-    {
-      where: {
-        CollectionId: collId
-      },
-      include: [taggerSchema.Tag],
-      attributes: ['"Tags.name"', 'id']
-    });
-
-};
-
 taggerDao.getCollectionsByArea = (areaId) => {
 
-  return taggerSchema.sequelize.query('Select * from Collections c LEFT JOIN AreaTargets at on c.id=at.CollectionId where at.AreaId = ? AND c.published = true order by c.title',
+  if (!areaId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
+  let areaArray = areaId.split(',');
+
+  let areaWhereClause = utils.getWhereClauseForMultipleAreas(areaArray);
+
+  return taggerSchema.sequelize.query('Select c.id, c.title, c.image, c.url, c.searchUrl, c.description, c.dates, c.items, c.browseType, c.repoType, c.restricted, c.published from Collections c LEFT JOIN AreaTargets at on c.id=at.CollectionId where ' + areaWhereClause + ' AND c.published = true group by c.id order by c.title',
     {
-      replacements: [areaId],
+      replacements: areaArray,
       type: taggerSchema.Sequelize.QueryTypes.SELECT
     });
 
@@ -379,17 +552,32 @@ taggerDao.getCollectionsByArea = (areaId) => {
 
 taggerDao.getCollectionsBySubjectAndArea = (subjectId, areaId) => {
 
-  return taggerSchema.sequelize.query('Select * from TagTargets tt LEFT JOIN Tags t on tt.TagId = t.id LEFT JOIN Collections c ' +
-    'on tt.CollectionId = c.id LEFT JOIN AreaTargets at on c.id=at.CollectionId where tt.TagId = ? and at.AreaId = ? and c.published = true ' +
-    'order by c.title',
+  if (!areaId || !subjectId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
+  let queryArray = areaId.split(',');
+
+  let areaWhereClause = utils.getWhereClauseForMultipleAreas(queryArray);
+
+  queryArray.unshift(subjectId);
+
+  return taggerSchema.sequelize.query('Select c.id, c.title, c.image, c.url, c.searchUrl, c.description, c.dates, c.items, c.browseType, c.repoType, c.restricted, c.published from TagTargets tt LEFT JOIN Tags t on tt.TagId = t.id LEFT JOIN Collections c ' +
+    'on tt.CollectionId = c.id LEFT JOIN AreaTargets at on c.id=at.CollectionId where tt.TagId = ? and ' + areaWhereClause + ' and c.published = true group by c.id order by c.title',
     {
-      replacements: [subjectId, areaId],
+      replacements: queryArray,
       type: taggerSchema.Sequelize.QueryTypes.SELECT
     });
 
 };
 
 taggerDao.getCollectionsBySubject = (subjectId) => {
+
+  if (!subjectId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
 
   return taggerSchema.sequelize.query('Select * from TagTargets tt LEFT JOIN Tags t on tt.TagId = t.id LEFT JOIN Collections c ' +
     'on tt.CollectionId = c.id where tt.TagId = ? and c.published = true order by c.title',
@@ -403,12 +591,50 @@ taggerDao.getCollectionsBySubject = (subjectId) => {
 
 taggerDao.getCollectionsByCategory = (categoryId) => {
 
+  if (!categoryId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
   return taggerSchema.sequelize.query('Select * from Collections c left join CategoryTargets ct on ct.CollectionId = c.id where ct.CategoryId = ? and c.published = true order by c.title',
     {
       replacements: [categoryId],
       type: taggerSchema.Sequelize.QueryTypes.SELECT
     });
 
+};
+
+taggerDao.findTagsForCollection = (collId) => {
+
+  if (!collId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
+
+  return taggerSchema.TagTarget.findAll(
+    {
+      where: {
+        CollectionId: collId
+      },
+      include: [taggerSchema.Tag],
+      attributes: ['TagId']
+    });
+
+};
+
+taggerDao.findRelatedCollections = (collId, subjectId) => {
+
+  if (!subjectId || !collId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
+  return taggerSchema.sequelize.query('Select c.title, c.id, c.image from Collections c left join TagTargets t on t.CollectionId = c.id where t.TagId = ? and t.CollectionId != ? and c.published = true order by c.id limit 6',
+    {
+      replacements: [subjectId, collId],
+      type: taggerSchema.Sequelize.QueryTypes.SELECT
+    });
 };
 
 module.exports = taggerDao;
