+++ server/api/tagger/dao/tags-dao.js
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016.
+ * Copyright (c) 2017.
  *
  *     This program is free software: you can redistribute it and/or modify
  *     it under the terms of the GNU General Public License as published by
@@ -21,8 +21,25 @@
 
 // jshint strict:false
 
-const taggerSchema = require('../models/index');
+const taggerSchema = require('../schema/index');
+const logger = require('../utils/error-logger');
+const path = require('path');
+const filename = path.basename(__filename);
+const utils = require('./utils');
+const paramErrorMessage = 'A parameter for a subject tag query is not defined.';
 const taggerDao = {};
+
+/**
+ * Returns 500 error for missing parameter. This error is thrown
+ * before the dao promise is returned.
+ * @returns {Error}
+ * @private
+ */
+function _errorResponse() {
+  let error = new Error('Error: missing query parameter - ' + filename);
+  error.status = 500;
+  return error;
+}
 
 taggerDao.findAllTags = () => {
 
@@ -35,6 +52,11 @@ taggerDao.findAllTags = () => {
 
 taggerDao.createTag = (name) => {
 
+  if(!name) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
   return taggerSchema.Tag.create({
     name: name
   });
@@ -42,6 +64,11 @@ taggerDao.createTag = (name) => {
 };
 
 taggerDao.updateTag = (name, id) => {
+
+  if(!name || !id) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
 
   return taggerSchema.Tag.update(
     {
@@ -56,6 +83,11 @@ taggerDao.updateTag = (name, id) => {
 
 taggerDao.deleteTag = (tagId) => {
 
+  if(!tagId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
   return taggerSchema.Tag.destroy({
     where: {
       id: tagId
@@ -65,6 +97,11 @@ taggerDao.deleteTag = (tagId) => {
 };
 
 taggerDao.findTagById = (tagId) => {
+
+  if(!tagId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
 
   return taggerSchema.Tag.find( {
     where: {
@@ -78,21 +115,44 @@ taggerDao.findTagById = (tagId) => {
 
 taggerDao.findTagsInArea = (areaId) => {
 
-  return taggerSchema.TagAreaTarget.findAll( {
-    where: {
-      AreaId: areaId
-    },
+  if(!areaId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
 
-    attributes: ['"Tags.name"', 'TagId'],
-    order: [[taggerSchema.Tag, 'name', 'ASC']],
-    include: [taggerSchema.Tag]
-  });
+  let areaArray = areaId.split(',');
+  let areaWhereClause = utils.getWhereClauseForMultipleAreas(areaArray);
+
+  return taggerSchema.sequelize.query('Select t.id, t.name from TagAreaTargets at LEFT JOIN Tags t on at.TagId = t.id  where ' + areaWhereClause + ' group by t.id order by t.name',
+    {
+      replacements: areaArray,
+      type: taggerSchema.Sequelize.QueryTypes.SELECT
+    });
+
+  // return taggerSchema.TagAreaTarget.findAll( {
+  //   where: {
+  //     AreaId: {
+  //       $or: [
+  //         areaArray
+  //       ]
+  //     }
+  //   },
+  //   attributes: ['TagId', '"Tag.name"'],
+  //   order: [[taggerSchema.Tag, 'name', 'ASC']],
+  //   group: [['TagId']],
+  //   include: [{ model: taggerSchema.Tag, attributes: id, name}]
+  // });
 
 };
 
 
 
 taggerDao.getTagCountByArea = (areaId) => {
+
+  if(!areaId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
 
   return taggerSchema.sequelize.query('SELECT name, COUNT(*) as count from TagTargets left join Tags on ' +
     'TagTargets.TagId = Tags.id left join TagAreaTargets on TagAreaTargets.TagId = Tags.id  ' +
@@ -106,6 +166,11 @@ taggerDao.getTagCountByArea = (areaId) => {
 
 taggerDao.findTagByName = (name) => {
 
+  if(!name) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
   return taggerSchema.Tag.find(
     {
       where: {
@@ -113,6 +178,24 @@ taggerDao.findTagByName = (name) => {
       }
     }
   );
+
+};
+
+taggerDao.findTagsForCollection = (collId) => {
+
+  if(!collId) {
+    logger.dao(paramErrorMessage);
+    throw _errorResponse();
+  }
+
+  return taggerSchema.TagTarget.findAll(
+    {
+      where: {
+        CollectionId: collId
+      },
+      include: [taggerSchema.Tag],
+      attributes: ['"Tags.name"', 'id']
+    });
 
 };
